Feature/methodes

+++ microservices/product/models/Product.js
@@ -14,7 +14,7 @@ Product.init(
       allowNull: false,
     },
     price: {
-      type: DataTypes.DECIMAL(10, 3),
+      type: DataTypes.INTEGER,
       allowNull: false,
     },
     description: {
